+++ admin/taikhoan/taikhoan.php
@@ -21,6 +21,19 @@ if (isset($_SESSION['isLoginOK']) && !empty($_SESSION['isLoginOK'])):
 
 endif;
 ?>
+<?php
+// Kiểm tra xem có thông báo nào được lưu trữ trong session không
+if(isset($_SESSION['message'])) {
+    $message = $_SESSION['message'];
+    $status = $_SESSION['status'];
+    // Hiển thị thông báo trên cùng trang
+    echo "<div class='alert alert-$status text-center'>$message</div>";
+    // Xóa thông báo và trạng thái khỏi session
+    unset($_SESSION['message']);
+    unset($_SESSION['status']);
+}
+// Tiếp tục xử lý trang danh sách sản phẩm
+?>
 
 <body id="body-pd">
     <header class="header" id="header">
@@ -207,7 +220,7 @@ endif;
                                         <td><a
                                                 href="update_account.php?id=<?php echo $id ?>&id1=<?php echo $row['id_admin']; ?>"><i
                                                     class="link-dark bi bi-pencil-square"></i></a></td>
-                                        <td><a
+                                        <td><a onclick="return confirm('Bạn có muốn xóa không?'); "
                                                 href="delete_account.php?id=<?php echo $id ?>&id1=<?php echo $row['id_admin']; ?>"><i
                                                     class="link-dark bi bi-trash"></i></a></td>
                                     </tr>
